Moved builder into separate file
Also limited each crop pool to 100 images
